+++ packages/apollos-data-connector-postgres/src/index.js
@@ -1,0 +1,9 @@
+export {
+  defineModel,
+  configureModel,
+  sequelize,
+  sync,
+  PostgresDataSource,
+} from './postgres';
+
+export * as Comment from './comments';