+++ packages/apollos-ui-connected/src/NodeSingleConnected/index.js
@@ -11,6 +11,7 @@ import ContentParentFeedConnected from '../ContentParentFeedConnected';
 import ContentChildFeedConnected from '../ContentChildFeedConnected';
 import UpNextButtonConnected from '../UpNextButtonConnected';
 import NodeFeaturesConnected from '../NodeFeaturesConnected';
+import ScriptureNodeConnected from '../ScriptureNodeConnected';
 
 import GET_MEDIA from './getMedia';
 
@@ -24,6 +25,7 @@ const NodeSingleInner = ({ nodeId, ImageWrapperComponent }) => (
       ImageWrapperComponent={ImageWrapperComponent}
       nodeId={nodeId}
     />
+    <ScriptureNodeConnected nodeId={nodeId} />
     <NodeFeaturesConnected nodeId={nodeId} />
     <UpNextButtonConnected nodeId={nodeId} />
     <ContentParentFeedConnected nodeId={nodeId} />
@@ -36,7 +38,7 @@ NodeSingleInner.propTypes = {
   ImageWrapperComponent: PropTypes.any, // eslint-disable-line
 };
 
-const NodeSingleConnected = ({ nodeId }) => (
+const NodeSingleConnected = ({ nodeId, children }) => (
   <BackgroundView>
     <StretchyView>
       {({ Stretchy, ...scrollViewProps }) => (
@@ -45,14 +47,16 @@ const NodeSingleConnected = ({ nodeId }) => (
         </FlexedScrollView>
       )}
     </StretchyView>
+    {children}
   </BackgroundView>
 );
 
 NodeSingleConnected.propTypes = {
   nodeId: PropTypes.string,
+  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 };
 
-const NodeSingleConnectedWithMedia = ({ nodeId }) => (
+const NodeSingleConnectedWithMedia = ({ nodeId, children }) => (
   <Query
     query={GET_MEDIA}
     variables={{ nodeId }}
@@ -60,18 +64,22 @@ const NodeSingleConnectedWithMedia = ({ nodeId }) => (
   >
     {({ data = {} }) => {
       if (!data?.node?.videos?.length)
-        return <NodeSingleConnected nodeId={nodeId} />;
-
+        return (
+          <NodeSingleConnected nodeId={nodeId}>{children}</NodeSingleConnected>
+        );
       return (
-        <ApollosPlayerContainer
-          source={data.node?.videos[0]?.sources[0]}
-          coverImage={data.node?.coverImage?.sources}
-          presentationProps={{
-            title: data.node.title,
-          }}
-        >
-          <NodeSingleInner nodeId={nodeId} ImageWrapperComponent={Noop} />
-        </ApollosPlayerContainer>
+        <>
+          <ApollosPlayerContainer
+            source={data.node?.videos[0]?.sources[0]}
+            coverImage={data.node?.coverImage?.sources}
+            presentationProps={{
+              title: data.node.title,
+            }}
+          >
+            <NodeSingleInner nodeId={nodeId} ImageWrapperComponent={Noop} />
+          </ApollosPlayerContainer>
+          {children}
+        </>
       );
     }}
   </Query>
@@ -79,6 +87,7 @@ const NodeSingleConnectedWithMedia = ({ nodeId }) => (
 
 NodeSingleConnectedWithMedia.propTypes = {
   nodeId: PropTypes.string,
+  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 };
 
 export default NodeSingleConnectedWithMedia;